Integrate wrangler
Data splitting is now done by the spark service instead of by
the pipeline. Logic from the splitter sink was moved to a
separate class, and the app now instantiates wrangler as a plugin
to parse and transform source file contents.

Right now, the wrangler transform itself is used. This should be
seen as a temporary solution, as we don't really want to include
cdap-etl-api in this app. However, wrangler-api currently
depends on it, so there must be some changes upstream before
anything can be done.