CDAP-13755 CDAP-13824 fix predictions dataset
If the outcome was not the first column in the training data, the
predictions dataset would shift columns over by one. This is because
the trainer was ordering the fields differently than the schema.
Fixed it so that the trainer follows the same schema as everything
else.

Also fixed a bug where the prediction for classifiers was not getting
reverse mapped before saving the predcitions. This would result in
unhelpful numeric labels in the predictions dataset instead of the
original string or boolean values.